change boost.application from master branch to version 4.11

+++ src/third-parties/boost-external/boost/application/aspects/args.hpp
@@ -58,7 +58,7 @@ namespace boost { namespace application {
        */
       int argc() const
       {
-         return (int)arguments_.size();
+         return arguments_.size();
       }
 
       /*!
